Test importing Glean.js in random file
Note that it was necessary to update webpack otherwise when building it
ignores Glean's `exports` statements.

Also, for now the `web-ext-types` needs to be included as a peer
dependency otherwise Glean.js will complain about missing types. It also
needs to be added to the typeRoots in Typescript's configuraiton. We
have a bug to make this unnecessary in the future.